seungso-mvp/backend/requirements.txt
# API & 서버
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings

# DB & 인프라 (Supabase 연동)
supabase==2.27.2
python-dotenv

# 데이터 처리 및 판결문 파싱
pandas==2.3.3
numpy==2.2.6
pymupdf==1.26.7
konlpy

# AI/ML 및 강화학습 (RL)
torch
scikit-learn
stable-baselines3[extra]
gymnasium

# 기타 유틸리티
httpx==0.28.1
python-dateutil==2.9.0.post0